requirements.txt
spotipy==2.24.0
SpeechRecognition==3.10.4
PyAudio==0.2.14
pyttsx3==2.90
python-dotenv==1.0.1